+++ examples/vite-app/src/examples/InteractiveMinimap/index.tsx
@@ -1,4 +1,4 @@
-import React, { MouseEvent } from 'react';
+import { MouseEvent, useCallback } from 'react';
 import ReactFlow, {
   MiniMap,
   Background,
@@ -8,6 +8,7 @@ import ReactFlow, {
   Node,
   Edge,
   useReactFlow,
+  XYPosition,
 } from 'reactflow';
 
 const onNodeDrag = (_: MouseEvent, node: Node) => console.log('drag', node);
@@ -17,35 +18,70 @@ const onNodeClick = (_: MouseEvent, node: Node) => console.log('click', node);
 const initialNodes: Node[] = [
   {
     id: '1',
-    type: 'input',
     data: { label: 'Node 1' },
-    position: { x: 250, y: 5 },
-    className: 'light',
+    position: { x: 0, y: 0 },
   },
   {
     id: '2',
     data: { label: 'Node 2' },
-    position: { x: 100, y: 100 },
-    className: 'light',
+    position: { x: 0, y: 200 },
   },
   {
     id: '3',
     data: { label: 'Node 3' },
-    position: { x: 400, y: 100 },
-    className: 'light',
+    position: { x: 200, y: 0 },
   },
+
   {
     id: '4',
     data: { label: 'Node 4' },
-    position: { x: 400, y: 200 },
-    className: 'light',
+    position: { x: 1000, y: 0 },
+  },
+  {
+    id: '5',
+    data: { label: 'Node 5' },
+    position: { x: 1000, y: 200 },
+  },
+  {
+    id: '6',
+    data: { label: 'Node 6' },
+    position: { x: 800, y: 0 },
+  },
+
+  {
+    id: '7',
+    data: { label: 'Node 4' },
+    position: { x: 0, y: 1000 },
+  },
+  {
+    id: '8',
+    data: { label: 'Node 5' },
+    position: { x: 0, y: 800 },
+  },
+  {
+    id: '9',
+    data: { label: 'Node 6' },
+    position: { x: 200, y: 1000 },
+  },
+
+  {
+    id: '10',
+    data: { label: 'Node 4' },
+    position: { x: 1000, y: 1000 },
+  },
+  {
+    id: '11',
+    data: { label: 'Node 5' },
+    position: { x: 800, y: 1000 },
+  },
+  {
+    id: '12',
+    data: { label: 'Node 6' },
+    position: { x: 1000, y: 800 },
   },
 ];
 
-const initialEdges: Edge[] = [
-  { id: 'e1-2', source: '1', target: '2', animated: true },
-  { id: 'e1-3', source: '1', target: '3' },
-];
+const initialEdges: Edge[] = [];
 
 const defaultEdgeOptions = { zIndex: 0 };
 
@@ -78,6 +114,14 @@ const BasicFlow = () => {
     );
   };
 
+  const onMiniMapClick = useCallback((event: MouseEvent, pos: XYPosition) => {
+    console.log(pos);
+  }, []);
+
+  const onMiniMapNodeClick = useCallback((event: MouseEvent, node: Node) => {
+    console.log(node);
+  }, []);
+
   return (
     <ReactFlow
       defaultNodes={initialNodes}
@@ -88,12 +132,12 @@ const BasicFlow = () => {
       className="react-flow-basic-example"
       minZoom={0.2}
       maxZoom={4}
-      fitView
       defaultEdgeOptions={defaultEdgeOptions}
       selectNodesOnDrag={false}
+      fitView
     >
       <Background variant={BackgroundVariant.Dots} />
-      <MiniMap />
+      <MiniMap onClick={onMiniMapClick} onNodeClick={onMiniMapNodeClick} pannable zoomable />
       <Controls />
 
       <div style={{ position: 'absolute', right: 10, top: 10, zIndex: 4 }}>
@@ -106,7 +150,7 @@ const BasicFlow = () => {
         <button onClick={toggleClassnames} style={{ marginRight: 5 }}>
           toggle classnames
         </button>
-        <button onClick={logToObject} style={{ marginRight: 5 }}>toObject</button>
+        <button onClick={logToObject}>toObject</button>
       </div>
     </ReactFlow>
   );